renamed models and other things to fix typo and continued views and templates work

+++ djanbs/jobs/views/views.py
@@ -1,10 +1,10 @@
 from django.contrib import messages
 from django.contrib.auth import login
-from django.shortcuts import redirect, render
+from django.shortcuts import Http404, redirect, render
 
-from jobs.forms import CandidactRegisterForm, CompanyRegisterForm, JobCandidactForm, JobOfferCreationForm
+from jobs.forms import CandidateRegisterForm, CompanyRegisterForm, JobCandidateForm, JobOfferCreationForm
 
-from ..models import Candidact, Company, JobCandidacted, JobOffer
+from ..models import Candidate, Company, JobCandidated, JobOffer
 from ..choices import EducationRequirement, PaymentRange, PositionLevel, int_to_string
 from ..decorators import allowed_groups
 
@@ -13,14 +13,20 @@ def home(request):
     if not request.user.is_authenticated:
         return redirect('login')
     else:
-        if request.user.groups.filter(name='candidact').exists():
+        if request.user.is_candidate:
             job_offers = JobOffer.objects.all() # type: ignore
             context = { 'job_offers': job_offers, }
             return render(request, 'jobs/index.html', context)
         else:
             company = Company.objects.get(pk=request.user.company.id) # type: ignore
             company_offers = JobOffer.objects.filter(company=company) # type: ignore
-            context = { 'company_offers': company_offers, }
+            job_candidated = [ 
+                JobCandidated.objects.filter( # type: ignore
+                job_offer=offer).values('candidate').distinct().count() 
+                for offer in company_offers ]
+            context = { 
+                'company_offers_cand': zip(company_offers,job_candidated),
+                }
             return render(request, 'jobs/company.html', context)
 
 
@@ -42,12 +48,11 @@ def create_job_offer(request, pk):
     return render(request, 'jobs/create-job.html', context)
 
 
-
 @allowed_groups(allowed_roles=['company'])
 def edit_job_offer(request, pk):
     job_offer = JobOffer.objects.get(id=pk) # type: ignore
     form = JobOfferCreationForm(instance=job_offer)
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user.company.id == job_offer.company.id:
         form = JobOfferCreationForm(request.POST, instance=job_offer)
         if form.is_valid():
             form.save()
@@ -62,7 +67,7 @@ def edit_job_offer(request, pk):
 @allowed_groups(allowed_roles=['company'])
 def delete_job_offer(request, pk):
     job_offer = JobOffer.objects.get(id=pk) # type: ignore
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user.company.id == job_offer.company.id:
         job_offer.delete()
         return redirect('home')
     context = {
@@ -71,49 +76,65 @@ def delete_job_offer(request, pk):
     return render(request, 'jobs/delete-job.html', context)
 
 
-@allowed_groups(allowed_roles=['candidact'])
-def candidact_to_job(request, job_id):
-    job_offer = JobOffer.objects.get(pk=job_id) # type: ignore
-    candidact = Candidact.objects.get(id=request.user.candidact.id) # type: ignore
+@allowed_groups(allowed_roles=['company'])
+def details_job_offer(request, pk):
+    job_offer = JobOffer.objects.get(id=pk) # type: ignore
+    if request.user.company.id == job_offer.company.id:
+        offer_candidates = [
+            jc for jc in JobCandidated.objects.filter(job_offer=job_offer)]
+        context = {
+            'job_offer': job_offer,
+            'offer_candidates': offer_candidates,
+            }
+        return render(request, 'jobs/details-job.html', context)
+    raise Http404
 
-    form = JobCandidactForm()
+
+@allowed_groups(allowed_roles=['candidate'])
+def candidate_to_job(request, job_id):
+    job_offer = JobOffer.objects.get(pk=job_id) # type: ignore
+    candidate = Candidate.objects.get(id=request.user.candidate.id) # type: ignore
+
+    form = JobCandidateForm()
     if request.method == 'POST':
-        form = JobCandidactForm(request.POST)
+        form = JobCandidateForm(request.POST)
         if form.is_valid():
             obj = form.save(commit=False)
             obj.job_offer = job_offer
-            obj.candidact = candidact
+            obj.candidate = candidate
             obj.save()
 
-            messages.success(request, f'Successfully candidacted to {job_offer}!')
+            messages.success(request, f'Successfully candidated to {job_offer}!')
             return redirect('home')
 
     context = {
         'form':form, 
         'job_offer': job_offer,
-        'candidact': candidact,
+        'candidate': candidate,
         'payment': int_to_string(PaymentRange, job_offer.payment_range),
         'education': int_to_string(EducationRequirement, job_offer.education_req),
         'pos_level': int_to_string(PositionLevel, job_offer.position_level),
     }
 
-    return render(request, 'jobs/job-candidact.html', context)
+    return render(request, 'jobs/job-candidate.html', context)
 
 
-def candidact_profile(request):
-    candidacted = JobCandidacted.objects.filter(candidact=request.user.candidact) # type: ignore
-    context = { 'candidacted': candidacted, }
-    return render(request, 'jobs/candidact_profile.html', context)
+@allowed_groups(allowed_roles=['candidate'])
+def candidate_profile(request):
+    candidated = JobCandidated.objects.filter(candidate=request.user.candidate) # type: ignore
+    context = { 'candidated': candidated, }
+    return render(request, 'jobs/candidate_profile.html', context)
 
 
+@allowed_groups(allowed_roles=['company'])
 def company_profile(request):
     context = {}
     return render(request, 'jobs/company_profile.html', context)
 
 
 def profile(request):
-    if request.user.groups.all()[0] == 'candidact':
-        return candidact_profile(request)
+    if request.user.is_candidate:
+        return candidate_profile(request)
     else:
         return company_profile(request)
 
@@ -133,20 +154,21 @@ def _edit_profile(request, form_model, template):
 
 
 def profile_edit(request):
-    if request.user.groups.filter(name='candidact').exists():
+    if request.user.is_candidate:
         return _edit_profile(request, 
-                CandidactRegisterForm, 'registration/cand-register.html')
+                CandidateRegisterForm, 'registration/cand-register.html')
     else:
         return _edit_profile(request, 
                 CompanyRegisterForm, 'registration/comp-register.html')
 
 
+@allowed_groups(allowed_roles=['candidate'])
 def delete_job_application(request, job_id):
-    job_candidacted = JobCandidacted.objects.get(pk=job_id) # type: ignore
+    job_candidated = JobCandidated.objects.get(pk=job_id) # type: ignore
     if request.method == "POST":
-        job_candidacted.delete()
+        job_candidated.delete()
         return redirect('profile')
     context = {
-        'job_candidacted': job_candidacted,
+        'job_candidated': job_candidated,
         }
     return render(request, 'jobs/delete-job-cand.html', context)
